--- electionsBR
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-import cepesp.api as electionsBr
-
-if __name__ == '__main__':  # pragma: no cover
-    echo "Ok"
